pyproject.toml
[tool.poetry]
name = "expenses-tracker"
version = "0.1.0"
description = "scan and track your expenses"
authors = ["gecBurton <g.e.c.burton@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
markitdown = {extras = ["pdf"], version = "^0.1.3"}
openai = "^1.54.3"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
